mxa-war/src/main/java/no/mxa/web/converter/YesNoConverter.java
package no.mxa.web.converter;

import javax.faces.component.UIComponent;
import javax.faces.context.FacesContext;
import javax.faces.convert.Converter;

import org.apache.commons.lang.NotImplementedException;

public class YesNoConverter implements Converter {

	@Override
	public Object getAsObject(FacesContext arg0, UIComponent arg1, String object) {
		// Not implemented
		throw new NotImplementedException();
	}

	@Override
	public String getAsString(FacesContext arg0, UIComponent arg1, Object object) {
		if (object.equals(Integer.valueOf(1))) {
			return "Ja";
		} else {
			return "Nei";
		}

	}

}
